Change terminal-table from under lib to the rubygem
