+++ lib/galileo.rb
@@ -3,7 +3,11 @@
 require "galileo/help"
 require "galileo/version"
 require "octokit"
-require "terminal-table/lib/terminal-table.rb"
+require "terminal-table"
+require "terminal-table/cell"
+require "terminal-table/row"
+require "terminal-table/style"
+require "terminal-table/table"
 require "time/time-ago-in-words"
 require "colorize"
 require "netrc"
