--- lib/terminal-table/Gemfile
@@ -1,7 +1,0 @@
-source 'http://rubygems.org'
-
-group :test do
-  gem 'echoe'
-  gem 'rspec', '1.3.1'
-  gem 'term-ansicolor', :require => "term/ansicolor"
-end
