--- lib/terminal-table/History.rdoc
@@ -1,48 +1,0 @@
-1.4.3 / 2011-10-13
-==================
-
-  * Optimize for faster table output.
-
-1.4.2 / 2010-01-14
-==================
-
-  * Fixed some bugs with colspan
-
-=== 1.4.1 / 2009-12-18
-
-* Fix column alignment with separators.
-
-=== 1.4.0 / 2009-12-18
-
-* Can now add :seperator arbitrarily in a table [thanks splattael]
-* Fix common typo: seperator -> separator [thanks splattael]
-
-=== 1.3.0 / 2009-10-16
-
-* Major refactoring (functionality remains the same)
-
-=== 1.2.0 / 2009-08-06
-
-* Added colspan support to table
-
-=== 1.1.0 / 2009-08-06
-
-* Added colspan support to table
-
-=== 1.1.0 / 2009-07-13
-
-* Added Table#==
-
-=== 1.0.5 / 2009-03-14
-
-* Allowing nil to be passed to table for headings
-* Revised doc to show that rows can be splatted now
-* Misc refactoring
-
-=== 1.0.3 / 2009-01-15
-
-* Moved yield or eval to Terminal::Table initialize where it belongs
-
-=== 1.0.0 / 2009-01-13
-
-* Initial release
