--- lib/terminal-table/Manifest
@@ -1,24 +1,0 @@
-History.rdoc
-Manifest
-README.rdoc
-Rakefile
-Todo.rdoc
-examples/examples.rb
-lib/terminal-table.rb
-lib/terminal-table/cell.rb
-lib/terminal-table/core_ext.rb
-lib/terminal-table/heading.rb
-lib/terminal-table/import.rb
-lib/terminal-table/table.rb
-lib/terminal-table/table_helper.rb
-lib/terminal-table/version.rb
-spec/cell_spec.rb
-spec/core_ext_spec.rb
-spec/import_spec.rb
-spec/spec.opts
-spec/spec_helper.rb
-spec/table_spec.rb
-tasks/docs.rake
-tasks/gemspec.rake
-tasks/spec.rake
-terminal-table.gemspec
