--- lib/terminal-table/Todo.rdoc
@@ -1,14 +1,0 @@
-
-== Major:
-
-* Nothing
-
-== Minor:
-
-* Programmatically add separator rows
-* Add multi-column sorting
-* Change; pre-create Cell and Heading objects to clean up Table a bit
-
-== Brainstorming:
-
-* Nothing
