--- lib/terminal-table/lib/terminal-table/core_ext.rb
@@ -1,8 +1,0 @@
-# encoding: utf-8
-class String
-  def align position, length
-    self.__send__ position, length
-  end
-  alias_method :left, :ljust
-  alias_method :right, :rjust
-end
