--- lib/terminal-table/lib/terminal-table/table_helper.rb
@@ -1,10 +1,0 @@
-# encoding: utf-8
-module Terminal
-  class Table
-    module TableHelper
-      def table headings = [], *rows, &block
-        Terminal::Table.new :headings => headings.to_a, :rows => rows, &block
-      end
-    end
-  end
-end
