--- lib/terminal-table/lib/terminal-table/version.rb
@@ -1,6 +1,0 @@
-# encoding: utf-8
-module Terminal
-  class Table
-    VERSION = '1.4.3'
-  end
-end
