+++ lib/terminal-table/row.rb
@@ -1,0 +1,10 @@
+module Terminal
+  class Table
+    class Row
+      def wrap_cell(index, width)
+        return if @cells.nil? or @cells[index].nil?
+        @cells[index].wrap(width)
+      end
+    end
+  end
+end
