--- lib/terminal-table/spec/spec.opts
@@ -1,1 +1,0 @@
---color
