--- lib/terminal-table/spec/spec_helper.rb
@@ -1,8 +1,0 @@
-
-require File.dirname(__FILE__) + '/../lib/terminal-table'
-
-class String
-  def deindent
-    strip.gsub(/^ */, '')
-  end
-end
