--- lib/terminal-table/tasks/docs.rake
@@ -1,13 +1,0 @@
-
-namespace :docs do
-  
-  desc 'Remove rdoc products'
-  task :remove => [:clobber_docs]
-  
-  desc 'Build docs, and open in browser for viewing (specify BROWSER)'
-  task :open => [:docs] do
-    browser = ENV["BROWSER"] || "safari"
-    sh "open -a #{browser} doc/index.html"
-  end
-  
-end
