--- lib/terminal-table/tasks/gemspec.rake
@@ -1,3 +1,0 @@
-
-desc 'Build gemspec file'
-task :gemspec => [:build_gemspec]
